requirements.txt
aiohttp==3.8.1
fastapi==0.73.0
loguru==0.6.0
prometheus-client==0.13.1
PyYAML==6.0
uvicorn==0.17.4
SQLAlchemy~=1.4.32
databases~=0.5.5
alembic~=1.7.7
asyncpg~=0.25.0
psycopg2-binary~=2.9.3